+++ test/unit/jasmine-boot.js
@@ -62,6 +62,7 @@ function initializePDFJS(callback) {
     'pdfjs-test/unit/evaluator_spec',
     'pdfjs-test/unit/fonts_spec',
     'pdfjs-test/unit/function_spec',
+    'pdfjs-test/unit/message_handler_spec',
     'pdfjs-test/unit/metadata_spec',
     'pdfjs-test/unit/murmurhash3_spec',
     'pdfjs-test/unit/network_spec',
@@ -74,7 +75,6 @@ function initializePDFJS(callback) {
     'pdfjs-test/unit/ui_utils_spec',
     'pdfjs-test/unit/unicode_spec',
     'pdfjs-test/unit/util_spec',
-    'pdfjs-test/unit/util_stream_spec',
   ].map(function (moduleName) {
     return SystemJS.import(moduleName);
   })).then(function(modules) {
